LayoutTests/fast/js/sputnik/Conformance/08_Types/8.6_The_Object_Type/8.6.2_Internal_Properties_and_Methods/S8.6.2_A7-expected.txt

S8.6.2_A7

PASS TypeError: Result of expression 'Math' [[object Math]] is not a constructor.

TEST COMPLETE
